+++ packages/core/odyssey-contribution-tooling/vite.config.project.ts
@@ -42,11 +42,18 @@ export default defineConfig(({ command }) => ({
     },
     sourcemap: true,
     rollupOptions: {
-      external: ["react", "react-dom"],
+      external: [
+        "react",
+        "react-dom",
+        "react/jsx-runtime",
+        "react/jsx-dev-runtime",
+      ],
       output: {
         globals: {
           react: "React",
           "react-dom": "ReactDOM",
+          "react/jsx-runtime": "ReactJSXRuntime",
+          "react/jsx-dev-runtime": "ReactJSXDevRuntime",
         },
       },
     },