+++ packages/core/odyssey-react-mui/src/test-selectors/index.ts
@@ -10,6 +10,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
+export * from "./linkedHtmlSelectors.js";
 export * from "./queryOdysseySelector.js";
 export * from "./querySelector.js";
 export * from "./testSelector.js";