+++ packages/core/odyssey-react-mui/src/ui-shell/SideNav/SideNavItemContentContext.tsx
@@ -13,6 +13,7 @@
 import { createContext, useContext } from "react";
 
 export type SideNavItemContentContextValue = {
+  absolutePaddingStart?: string;
   depth: number;
   isCompact?: boolean;
   isSortable?: boolean;