+++ packages/tools/odyssey-cli/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.47.0](https://github.com/atko-eng/odyssey-design-system/compare/v1.46.0...v1.47.0) (2025-10-28)
+
+**Note:** Version bump only for package @okta/odyssey-cli
+
 ## [1.46.0](https://github.com/atko-eng/odyssey-design-system/compare/v1.45.2...v1.46.0) (2025-10-16)
 
 **Note:** Version bump only for package @okta/odyssey-cli